feat: route registration
When using multi-tenant mode, ACA-Py uses a routing table mapping
recipient key to the recipient wallet.
Routes are created when invitations are created. If a third-party builds
an invitation from a public DID Document (so-called "implicit
invitation"), the verification key referenced in the DIDComm service
needs to be registered in the routing table.

This commit adds an endpoint allowing to register the verification key
of a DID in the routing table.